requirements.txt
Flask
gunicorn
pandas
-e .
Pyyaml
six
sklearn
jsonify
pickle